+++ src/anyio/abc/_eventloop.py
@@ -26,7 +26,7 @@ else:
 if TYPE_CHECKING:
     from typing import Literal
 
-    from .._core._synchronization import CapacityLimiter, Event
+    from .._core._synchronization import CapacityLimiter, Event, Lock
     from .._core._tasks import CancelScope
     from .._core._testing import TaskInfo
     from ..from_thread import BlockingPortal
@@ -165,6 +165,11 @@ class AsyncBackend(metaclass=ABCMeta):
     @classmethod
     @abstractmethod
     def create_event(cls) -> Event:
+        pass
+
+    @classmethod
+    @abstractmethod
+    def create_lock(cls) -> Lock:
         pass
 
     @classmethod